Try to fix resizing problem
I'm calling this the Izma solution: I put that terminal in a box, and I
put that box in a window and when the window is resized, I smash it!

The Gtk+ nesting is now:

Window -> Notebook -> Box => Terminal
                       |        |
                       \->TMux>-/
The Terminal is never resized beyond what TMux requests. When the Box is
enlarged, the Box tells TMux to make the client bigger, which, in turn,
makes the Terminal bigger. This also means if there is another client
that is smaller than this client, there won't be terminal corruption.